Changes to the color_shade_gen function intended for research work
Add to the step function in the form of a tuple to determine the frequency of creating different shades of colors. Remove rgb, because it makes no sense to create 1 picture and change the background color in it, when you can immediately create a picture with the desired background color.

Conclusion: modified function that will be able to generate 16 million images of different colors