Сreating a color_shade_gen that will generate images with different shades of color to test the check_text function on different background colors (#143)
* Fix linter R0917 error (too many positional arguments)

When execute the push, the code does not pass the github actions check, as error R0917 occurs related to positional arguments. To solve the problem, need to add a restriction on checking linter for a part of the code so that this does not apply to the entire API of the library

In conclusion, after check github actions, expect the error to be fixed

* Add color generator and try to generate pictures with different color fills

* Tested the check_test function using the first 2000 photos with a blue background, but with different shades of colors

* created new function 'background changer'

* add crop function

* add crop function

* add 2 pictures as example of bg_changer work

* clean up

* made an example of work bg_changer and check_text for future histogram

* add histogram and fix test

* clean up

* clean up

* Create a color shade generator and test check_text

Create the color_shade_gen function in order to investigate the dependence of the tesseract recognition quality on the background color and make a test for the check_text function, which uses a library with a built-in neural network

Built a histogram where display the results of text recognition in different shades of color

Conclusion: the accuracy of text recognition in different background colors depends on contrast

* clean up

* Add new folders to .gitignore

In order for files creat use color_shade_gen and words_for_bg fixtures not to get into github when use git add * , new entries add to .gitignore. This will help to avoid memory loss occupied by unnecessary data.

Conclusion: images with different background colors do not appear on github

* Made a docstring adjustment for color_shade_gen

Add an optional value to the size argument. if this argument is missing when declaring the function, the default value of 1920x1080 will be substituted

Conclusion: improved docstring description

* Changes to the color_shade_gen function intended for research work

Add to the step function in the form of a tuple to determine the frequency of creating different shades of colors. Remove rgb, because it makes no sense to create 1 picture and change the background color in it, when you can immediately create a picture with the desired background color.

Conclusion: modified function that will be able to generate 16 million images of different colors

* Final part of the code that will fix the error in github actions with ERROR collecting tests

The following error came out ' '

To solve the problem, need to update opencv-python==4.10.0.84, because numpy depends on it

Conclusion: github actions are being verified successfully

* Fix after code review

* reuse folder_path in save

* reuse variables in save

* reuse variables in save

* Stop hardcode path and filename replacing them with variables in the code

Replaced paths with variables, as this will help make the year more flexible in case the path to a specific file or path is changed

Conclusion: code has become more flexible to changes

* Modifie the color_shade_gen function

Modified the color_shade_gen function in such a way that now, with each iteration, RGB values change due to step. This improvement implements the correct iterator for colors.

Conclusion: the function works correctly and as intended

* Сlean up in the color_shade_gen function

In order not to create an initial image with the background color (0, 0, 0) outside the loop, the iteration process was modified.

Conclusion: the function works correctly and as intended

* Create a  color iterator for merge

Сreate a color iterator that creates an iterator type that iterates from the initial rgb to the specified limit using step, which changes the value of the color tuple each time

Conclusion: the correct iterator for colors has been implemented

* clean up

* rename class name in gen.py

* clean up

* add result of the work with less count of files

* fix .gitignore

* failed push. rework

* Delete sample images from the repository

All images will not be needed in the repository, because any user can generate these files. We only needed them for testing.

Conclusion: delete images from the repository

---------

Co-authored-by: Иван Горбунов <muervos@muervos.muervos>