Add a function to search for key points in the image
Text recognition will be based on key points. The key points are the interesting areas of the image. Interesting areas are those areas that are heterogeneous. For example, these can be angles, as there is a sharp change in intensity in two different directions

Conclusion: understand in which areas there is a sharp change in color intensity