Create a key point clustering function
Create a key point clustering function. This is necessary to find the area of concentration of points using K-means, the algorithm of which is based on the assumption that descriptors that are close to each other in the feature space represent similar key points. Thus, K-means searches for groups of descriptors that have similar meanings.

Conclusion:  group key points based on their descriptors