Create a function that compares two images
Create a function that compares two images by detecting and matching key points. This can be useful for comparing scada images for changes in the state of objects.

The feature is under development and will be finalized. One of the ideas is to evaluate the similarity of images, identify places where key points differ.

Conclusion: This function will help  analyze objects.